requirements.txt
awacs==0.5.4
troposphere==1.8.1
pyyaml
